+++ include/Simulation.h
@@ -6,13 +6,11 @@
 
 #include "Logger.h"
 #include "HDF5File.h"
+#include "ClosedLoopHDF5File.h"
 #include "TemperatureGenerator.h"
 #include "TemperatureFromFile.h"
 #include "NominalTemperature.h"
-#include "Detector.h"
-#include "DetectorWithMappedPSF.h"
-#include "DetectorWithAnalyticGaussianPSF.h"
-#include "DetectorWithAnalyticNonGaussianPSF.h"
+
 #include "Camera.h"
 #include "Telescope.h"
 #include "Platform.h"
@@ -21,12 +19,17 @@
 #include "NoJitter.h"
 #include "JitterFromFile.h"
 #include "JitterFromRedNoise.h"
+#include "JitterFromNetwork.h"
 #include "DriftGenerator.h"
 #include "NoDrift.h"
 #include "ThermoElasticDriftFromFile.h"
 #include "ThermoElasticDriftFromRedNoise.h"
 #include "ConfigurationParameters.h"
 #include "version.h"
+
+#include "AbstractDetectorFactory.h"
+#include "ClosedLoopUtility.h"
+#include "ClosedLoopDetectorClasses.h"
 
 
 using namespace std;
@@ -53,6 +56,7 @@ class Simulation
     private:
 
         double currentTime;
+        double cycleTime;
         double exposureTime;
         double readoutTimeBeforeNextExposure;	// Readout time before the next exposure starts
 
@@ -60,7 +64,7 @@ class Simulation
         int numExposures;                       // Number of exposures
 
         bool useJitter;
-        bool useJitterFromFile;
+        string jitterSource;
         bool includeFieldDistortion;
         bool useDrift;
         bool useDriftFromFile;
@@ -80,7 +84,12 @@ class Simulation
         Camera *camera;
         Detector *detector;
 
-        HDF5File hdf5File;
+        HDF5File *hdf5File;
+
+        AbstractDetectorFactory* detectorFactory;
+
+        bool sendImagettesToClient;
+        bool getWindowPositionFromServer;
 
 };